Moves responses to the right path and adds model
