+++ app/src/main/java/io/github/koss/mammut/feature/instancebrowser/about/InstanceAboutViewModel.kt
@@ -3,7 +3,7 @@ package io.github.koss.mammut.feature.instancebrowser.about
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import io.github.koss.mammut.instances.InstanceDetail
+import io.github.koss.mammut.instances.response.InstanceDetail
 import io.github.koss.mammut.extension.postSafely
 
 class InstanceAboutViewModel: ViewModel() {
