+++ app/src/main/java/io/github/koss/mammut/feature/instancebrowser/recyclerview/InstanceCardViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import io.github.koss.mammut.data.models.InstanceRegistration
-import io.github.koss.mammut.instances.InstanceDetail
+import io.github.koss.mammut.instances.response.InstanceDetail
 import io.github.koss.mammut.repo.InstancesRepository
 import io.github.koss.mammut.extension.postSafely
 import kotlinx.coroutines.CoroutineScope
