+++ app/src/main/java/io/github/koss/mammut/repo/InstancesRepository.kt
@@ -7,7 +7,7 @@ import io.github.koss.mammut.data.database.MammutDatabase
 import io.github.koss.mammut.data.database.entities.InstanceSearchResultEntity
 import io.github.koss.mammut.data.models.InstanceSearchResult
 import io.github.koss.mammut.instances.MastodonInstancesService
-import io.github.koss.mammut.instances.InstanceDetail
+import io.github.koss.mammut.instances.response.InstanceDetail
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
