+++ toot/src/main/java/io/github/koss/mammut/toot/TootModel.kt
@@ -1,0 +1,18 @@
+package io.github.koss.mammut.toot
+
+import android.os.Parcelable
+import com.sys1yagi.mastodon4j.api.entity.Status
+import kotlinx.android.parcel.Parcelize
+
+/**
+ * Model containing the required information for submitting a new [Status]
+ */
+@Parcelize
+data class TootModel(
+        val status: String,
+        val inReplyToId: Long?,
+        val mediaIds: List<Long>?,
+        val sensitive: Boolean,
+        val spoilerText: String?,
+        val visibility: Status.Visibility = Status.Visibility.Public
+): Parcelable
